Refactor: committer inside consensus now and server moved to a package.
