+++ protocol/leaderrotation/carousel/carousel.go
@@ -9,10 +9,10 @@ import (
 	"github.com/relab/hotstuff/core"
 	"github.com/relab/hotstuff/core/logging"
 	"github.com/relab/hotstuff/modules"
+	"github.com/relab/hotstuff/protocol/committer"
 	"github.com/relab/hotstuff/protocol/leaderrotation"
 	"github.com/relab/hotstuff/protocol/synchronizer/viewduration"
 	"github.com/relab/hotstuff/security/blockchain"
-	"github.com/relab/hotstuff/service/committer"
 )
 
 const ModuleName = "carousel"
