+++ wiring/consensus.go
@@ -6,9 +6,9 @@ import (
 	"github.com/relab/hotstuff/core/logging"
 	"github.com/relab/hotstuff/internal/proto/clientpb"
 	"github.com/relab/hotstuff/modules"
+	"github.com/relab/hotstuff/protocol/committer"
 	"github.com/relab/hotstuff/protocol/consensus"
 	"github.com/relab/hotstuff/security/cert"
-	"github.com/relab/hotstuff/service/committer"
 )
 
 type Consensus struct {
